+++ src/providers/completionItemProvider.ts
@@ -10,6 +10,7 @@ import { ExtensionContext, TextDocument, Position, CompletionItem } from 'vscode
 import { loadTable } from '../extension';
 import { Parser } from '../language/parser';
 import { Modules } from '../language/modules';
+import { Language } from '../language/language';
 
 class AttributeItem{
     public attribute:string = '';
@@ -44,6 +45,40 @@ export class TemplateCompletionItemProvider implements vscode.CompletionItemProv
                 }
             }
 
+            // seek in language tables and imported modules
+            let items: CompletionItem[] = [];
+            let prefix = line.text.substring(line.text.length - 1);
+            if (line.text[line.text.length - 2] === '%') {
+                for (const symbol of Language.getInstance().findSymbols('%' + prefix)) {
+                    let item = new vscode.CompletionItem(symbol.signature, vscode.CompletionItemKind.Property);
+                    item.detail = symbol.signature;
+                    item.documentation = symbol.summary;
+                    items.push(item);
+                }
+                return resolve(items);
+            }
+            for (const languageItem of Language.getInstance().seekByPrefix(prefix)) {
+                let prettySignature = Language.prettySignature(languageItem.signature);
+                let item = new vscode.CompletionItem(prettySignature, TemplateCompletionItemProvider.getCompletionItemKind(languageItem.category));
+                item.insertText = new vscode.SnippetString(languageItem.signature);
+                item.detail = prettySignature;
+                item.documentation = languageItem.summary;
+                items.push(item);
+            }
+            for (const result of Modules.getInstance().findActions(prefix)) {
+                for (const overload of result.action.overloads) {
+                    let signature = Modules.prettySignature(overload);
+                    let item = new vscode.CompletionItem(signature, TemplateCompletionItemProvider.getCompletionItemKind(result.actionKind));
+                    item.insertText = new vscode.SnippetString(overload);
+                    item.detail = result.moduleName + ' -> ' + signature;
+                    item.documentation = result.action.summary;
+                    items.push(item);
+                }
+            }
+            if (items.length > 0) {
+                return resolve(items);
+            }
+
             // quests
             if (Parser.isQuest(document.lineAt(position.line).text)) {
                 return Parser.findLinesInAllfiles(Parser.questDefinitionPattern, true).then(results => {
@@ -70,24 +105,22 @@ export class TemplateCompletionItemProvider implements vscode.CompletionItemProv
                 }, () => { return reject(); });
             }
 
-            // actions
-            let items: CompletionItem[] = [];
-            for (const result of Modules.getInstance().findActions(line.text.substring(line.text.length - 1))) {
-                for (const overload of result.action.overloads) {
-                    let signature = Modules.prettySignature(overload);
-                    let kind = result.actionKind === Modules.ActionKind.Action ? vscode.CompletionItemKind.Method : vscode.CompletionItemKind.Event;
-                    let item = new vscode.CompletionItem(signature, kind);
-                    item.insertText = new vscode.SnippetString(overload);
-                    item.detail = result.moduleName + ' -> ' + signature;
-                    item.documentation = result.action.summary;
-                    items.push(item);
-                }
-            }
-            if (items.length > 0) {
-                return resolve(items);
-            }
-
             return reject();
         });
+    }
+
+    private static getCompletionItemKind(category: string): vscode.CompletionItemKind {
+        switch (category) {
+            case Language.ItemKind.Keyword:
+                return vscode.CompletionItemKind.Keyword;
+            case Language.ItemKind.Message:
+                return vscode.CompletionItemKind.Struct;
+            case Modules.ActionKind.Action:
+                return vscode.CompletionItemKind.Method;
+            case Modules.ActionKind.Condition:
+                return vscode.CompletionItemKind.Event;
+            default:
+                return vscode.CompletionItemKind.Text;
+        }
     }
 }